refactor(ts): only generate tsconfig.json if missing (#5356) (#5367)
