+++ packages/cli/src/utils/index.js
@@ -101,6 +101,10 @@ export function showBanner(nuxt) {
   // Running mode
   titleLines.push(`Running in ${nuxt.options.dev ? chalk.bold.blue('development') : chalk.bold.green('production')} mode (${chalk.bold(nuxt.options.mode)})`)
 
+  if (nuxt.options._typescript) {
+    titleLines.push(`TypeScript support is ${chalk.green.bold('enabled')}`)
+  }
+
   // https://nodejs.org/api/process.html#process_process_memoryusage
   const { heapUsed, rss } = process.memoryUsage()
   titleLines.push(`Memory usage: ${chalk.bold(prettyBytes(heapUsed))} (RSS: ${prettyBytes(rss)})`)
