+++ test/fixtures/basic/pages/jsx-link.js
@@ -1,4 +1,4 @@
-import renderLink from './jsx-link.jsx'
+import renderLink from './jsx-link2.jsx'
 
 export default {
   render: renderLink
