+++ test/fixtures/basic/pages/jsx-link2.jsx
